Arrange functions to live on proper pages
