+++ src/UserRepository.js
@@ -1,19 +1,20 @@
 import sleepData from './data/sleep';
 
 class UserRepository {
-  constructor() {
-    this.users = [];
+  constructor(userData) {
+    this.users = userData;
   }
+
   getUser(id) {
-    return this.users.find(function(user) {
+    return this.users.find(function (user) {
       return user.id === id;
-    })
+    });
   }
   calculateAverageStepGoal() {
-    let goals = this.users.map(function(user) {
+    let goals = this.users.map(function (user) {
       return user.dailyStepGoal;
     });
-    let total = goals.reduce(function(sum, goal) {
+    let total = goals.reduce(function (sum, goal) {
       sum += goal;
       return sum;
     }, 0);
@@ -27,74 +28,87 @@ class UserRepository {
     return totalSleepQuality / this.users.length;
   }
   calculateAverageSteps(date) {
-    let allUsersStepsCount = this.users.map(user => {
-      return user.activityRecord.filter(activity => {
+    let allUsersStepsCount = this.users.map((user) => {
+      return user.activityRecord.filter((activity) => {
         return activity.date === date;
       });
-    })
-    let sumOfSteps = allUsersStepsCount.reduce((stepsSum, activityCollection) => {
-      activityCollection.forEach(activity => {
-        stepsSum += activity.steps
-      })
-      return stepsSum;
-    }, 0);
+    });
+    let sumOfSteps = allUsersStepsCount.reduce(
+      (stepsSum, activityCollection) => {
+        activityCollection.forEach((activity) => {
+          stepsSum += activity.steps;
+        });
+        return stepsSum;
+      },
+      0
+    );
     return Math.round(sumOfSteps / allUsersStepsCount.length);
   }
   calculateAverageStairs(date) {
-    let allUsersStairsCount = this.users.map(user => {
-      return user.activityRecord.filter(activity => {
+    let allUsersStairsCount = this.users.map((user) => {
+      return user.activityRecord.filter((activity) => {
         return activity.date === date;
       });
-    })
-    let sumOfStairs = allUsersStairsCount.reduce((stairsSum, activityCollection) => {
-      activityCollection.forEach(activity => {
-        stairsSum += activity.flightsOfStairs
-      })
-      return stairsSum;
-    }, 0);
+    });
+    let sumOfStairs = allUsersStairsCount.reduce(
+      (stairsSum, activityCollection) => {
+        activityCollection.forEach((activity) => {
+          stairsSum += activity.flightsOfStairs;
+        });
+        return stairsSum;
+      },
+      0
+    );
     return Math.round(sumOfStairs / allUsersStairsCount.length);
   }
   calculateAverageMinutesActive(date) {
-    let allUsersMinutesActiveCount = this.users.map(user => {
-      return user.activityRecord.filter(activity => {
+    let allUsersMinutesActiveCount = this.users.map((user) => {
+      return user.activityRecord.filter((activity) => {
         return activity.date === date;
       });
-    })
-    let sumOfMinutesActive = allUsersMinutesActiveCount.reduce((minutesActiveSum, activityCollection) => {
-      activityCollection.forEach(activity => {
-        minutesActiveSum += activity.minutesActive
-      })
-      return minutesActiveSum;
-    }, 0);
+    });
+    let sumOfMinutesActive = allUsersMinutesActiveCount.reduce(
+      (minutesActiveSum, activityCollection) => {
+        activityCollection.forEach((activity) => {
+          minutesActiveSum += activity.minutesActive;
+        });
+        return minutesActiveSum;
+      },
+      0
+    );
     return Math.round(sumOfMinutesActive / allUsersMinutesActiveCount.length);
   }
   calculateAverageDailyWater(date) {
-    let todaysDrinkers = this.users.filter(user => {
+    let todaysDrinkers = this.users.filter((user) => {
       return user.addDailyOunces(date) > 0;
     });
     let sumDrankOnDate = todaysDrinkers.reduce((sum, drinker) => {
-      return sum += drinker.addDailyOunces(date);
-    }, 0)
+      return (sum += drinker.addDailyOunces(date));
+    }, 0);
     return Math.floor(sumDrankOnDate / todaysDrinkers.length);
   }
   findBestSleepers(date) {
-    return this.users.filter(user => {
+    return this.users.filter((user) => {
       return user.calculateAverageQualityThisWeek(date) > 3;
-    })
+    });
   }
   getLongestSleepers(date) {
-    return sleepData.filter(sleep => {
-      return sleep.date === date;
-    }).sort((a, b) => {
-      return b.hoursSlept - a.hoursSlept;
-    })[0].userID;
+    return sleepData
+      .filter((sleep) => {
+        return sleep.date === date;
+      })
+      .sort((a, b) => {
+        return b.hoursSlept - a.hoursSlept;
+      })[0].userID;
   }
   getWorstSleepers(date) {
-    return sleepData.filter(sleep => {
-      return sleep.date === date;
-    }).sort((a, b) => {
-      return a.hoursSlept - b.hoursSlept;
-    })[0].userID;
+    return sleepData
+      .filter((sleep) => {
+        return sleep.date === date;
+      })
+      .sort((a, b) => {
+        return a.hoursSlept - b.hoursSlept;
+      })[0].userID;
   }
 }
